# app/Http/Requests/UpdateInfoRequest.php
<?php

namespace App\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;

/**
 * @OA\Schema (
 *     title="Update user info request",
 *     description="Update user info request body data",
 * )
 */
class UpdateInfoRequest extends FormRequest
{
    /**
     * @OA\Property (
     *     title="fist_name"
     * )
     * @var string
     */
    public $first_name;

    /**
     * @OA\Property (
     *     title="last_name"
     * )
     * @var string
     */
    public $last_name;

    /**
     * @OA\Property (
     *     title="email"
     * )
     * @var string
     */
    public $email;
    /**
     * Determine if the user is authorized to make this request.
     */
    public function authorize(): bool
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
     */
    public function rules(): array
    {
        return [
            'email' => 'email|unique:users',
        ];
    }
}