# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "description": "Simple role based authentication API Documentation in Laravel 11.x",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "admin@laravel-api.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Admin API Service"
        }
    ],
    "paths": {
        "/user": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "operationId": "8f60355fe45caf090eec729e26434da4",
                "responses": {
                    "200": {
                        "description": "Authenticated User"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/info": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "operationId": "06958ee4e97120f131a2bfd9657165cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateInfoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "User Info Update"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/password": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "operationId": "d8ea7d5c62a91478d6ea91a5816713fa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "User Password Update"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "operationId": "7208c29037df2b8bb02499ba39ed8a78",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Pagination Page",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users Collection"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "operationId": "262eb5ca1fdea354023c0a8870442c1e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User Create "
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "operationId": "6946bbab7898f59a09e3c8e529a38907",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Uesr ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Show"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "operationId": "05028d02ea3e9a99abd6260cfe05b179",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "User Update"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "operationId": "b3d28da5dc9de7fc1bd425f93b54ff7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Uesr ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User Delete"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "UpdateInfoRequest": {
                "title": "Update user info request",
                "description": "Update user info request body data",
                "properties": {
                    "first_name": {
                        "title": "fist_name",
                        "type": "string"
                    },
                    "last_name": {
                        "title": "last_name",
                        "type": "string"
                    },
                    "email": {
                        "title": "email",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdatePasswordRequest": {
                "title": "Update user password request",
                "description": "Update user password request body data",
                "properties": {
                    "password": {
                        "title": "password",
                        "type": "string"
                    },
                    "password_confirm": {
                        "title": "password_confirm",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateUserRequest": {
                "title": "Update user request",
                "description": "Update user request body data",
                "properties": {
                    "first_name": {
                        "title": "fist_name",
                        "type": "string"
                    },
                    "last_name": {
                        "title": "last_name",
                        "type": "string"
                    },
                    "email": {
                        "title": "email",
                        "type": "string"
                    },
                    "role_id": {
                        "title": "role_id",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "UserCreateRequest": {
                "title": "Store user request",
                "description": "Store user request body data",
                "properties": {
                    "first_name": {
                        "title": "fist_name",
                        "type": "string"
                    },
                    "last_name": {
                        "title": "last_name",
                        "type": "string"
                    },
                    "email": {
                        "title": "email",
                        "type": "string"
                    },
                    "role_id": {
                        "title": "role_id",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Profile",
            "description": "Profile"
        },
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}